requirements_stable.txt
# Ultra stable requirements for bot deployment
python-telegram-bot==20.0a2
python-dotenv==1.0.0